admin fixes 1

+++ app/Http/Controllers/Admin/EmployeeRoleController.php
@@ -8,6 +8,7 @@ use App\Http\Requests\UpdateEmployeeRoleRequest;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Auth;
 
 class EmployeeRoleController extends Controller
 {
@@ -18,20 +19,25 @@ class EmployeeRoleController extends Controller
      */
     public function index()
     {
-        session(
-            [
-                'menu_active' => 'designations',
-                'page_title' => 'Designations',
-                'page_title_description' => 'Manage Designations & Details',
-                'breadcrumb' => [
-                    'Home' => route('admin.dashboard'),
-                    'Designations' => route('designations.index'),
-                ],
-            ]
-        );
+        if(Auth::check() == false){
+            return redirect()->route('login');
+        }
+        else{
+            session(
+                [
+                    'menu_active' => 'designations',
+                    'page_title' => 'Designations',
+                    'page_title_description' => 'Manage Designations & Details',
+                    'breadcrumb' => [
+                        'Home' => route('admin.dashboard'),
+                        'Designations' => route('designations.index'),
+                    ],
+                ]
+            );
 
-        $employeeRoles = EmployeeRole::all();
-        return view('admin.employees.emp-roles', compact('employeeRoles'));
+            $employeeRoles = EmployeeRole::all();
+            return view('admin.employees.emp-roles', compact('employeeRoles'));
+        }
     }
 
     /**
@@ -52,20 +58,25 @@ class EmployeeRoleController extends Controller
      */
     public function store(StoreEmployeeRoleRequest $request)
     {
-        $validated_data = Validator::make($request->all(), $request->rules(), $request->messages());
-
-        if ($validated_data->fails()) {
-            return redirect()->back()->withErrors($validated_data->errors());
+        if(Auth::check() == false){
+            return redirect()->route('login');
         }
+        else{
+            $validated_data = Validator::make($request->all(), $request->rules(), $request->messages());
 
-        $employeeRole = new EmployeeRole();
-        $employeeRole->name = $request->name;
-        $employeeRole->description = $request->description;
-        $employeeRole->status = $request->status;
-        $employeeRole->slug = Str::slug($request->name);
-        $employeeRole->save();
+            if ($validated_data->fails()) {
+                return redirect()->back()->withErrors($validated_data->errors());
+            }
 
-        return redirect()->back()->with('success', 'Employee Role created successfully');
+            $employeeRole = new EmployeeRole();
+            $employeeRole->name = $request->name;
+            $employeeRole->description = $request->description;
+            $employeeRole->status = $request->status;
+            $employeeRole->slug = Str::slug($request->name);
+            $employeeRole->save();
+
+            return redirect()->back()->with('success', 'Employee Role created successfully');
+        }
     }
 
     /**
@@ -99,23 +110,28 @@ class EmployeeRoleController extends Controller
      */
     public function update(UpdateEmployeeRoleRequest $request, $employeeRole)
     {
-        $validated_data = Validator::make($request->all(), $request->rules(), $request->messages());
-        $employeeRole = EmployeeRole::find($employeeRole);
-
-        if ($validated_data->fails()) {
-            return redirect()->back()->withErrors($validated_data->errors());
+        if(Auth::check() == false){
+            return redirect()->route('login');
         }
-        else if (!$employeeRole) {
-            return redirect()->back()->withErrors('Employee Role not found');
+        else{
+            $validated_data = Validator::make($request->all(), $request->rules(), $request->messages());
+            $employeeRole = EmployeeRole::find($employeeRole);
+
+            if ($validated_data->fails()) {
+                return redirect()->back()->withErrors($validated_data->errors());
+            }
+            else if (!$employeeRole) {
+                return redirect()->back()->withErrors('Employee Role not found');
+            }
+
+            $employeeRole->name = $request->name;
+            $employeeRole->description = $request->description;
+            $employeeRole->status = $request->status;
+            $employeeRole->slug = Str::slug($request->name);
+            $employeeRole->update();
+
+            return redirect()->back()->with('success', 'Employee Role updated successfully');
         }
-
-        $employeeRole->name = $request->name;
-        $employeeRole->description = $request->description;
-        $employeeRole->status = $request->status;
-        $employeeRole->slug = Str::slug($request->name);
-        $employeeRole->update();
-
-        return redirect()->back()->with('success', 'Employee Role updated successfully');
     }
 
     /**
@@ -126,17 +142,22 @@ class EmployeeRoleController extends Controller
      */
     public function destroy($employeeRole)
     {
-        $employeeRole = EmployeeRole::find($employeeRole);
-
-        if (!$employeeRole) {
-            return redirect()->back()->withErrors('Employee Role not found');
+        if(Auth::check() == false){
+            return redirect()->route('login');
         }
-        else if ($employeeRole->employees()->count() > 0) {
-            return redirect()->back()->withErrors('Employee Role cannot be deleted because it is assigned to one or more employees');
+        else{
+            $employeeRole = EmployeeRole::find($employeeRole);
+
+            if (!$employeeRole) {
+                return redirect()->back()->withErrors('Employee Role not found');
+            }
+            else if ($employeeRole->employees()->count() > 0) {
+                return redirect()->back()->withErrors('Employee Role cannot be deleted because it is assigned to one or more employees');
+            }
+
+            $employeeRole->delete();
+
+            return redirect()->back()->with('success', 'Employee Role deleted successfully');
         }
-
-        $employeeRole->delete();
-
-        return redirect()->back()->with('success', 'Employee Role deleted successfully');
     }
 }
